Consolidate recipe editor actions into one file
