+++ websites/recipe-website/editor/src/actions/updateRecipe/index.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { auth, signIn } from "@/auth";
-import updateRecipe from "recipe-editor/controller/actions/update";
+import { updateRecipe } from "recipe-editor/controller/actions";
 import { RecipeFormState } from "recipe-website-common/controller/formState";
 
 export default async function authenticateAndUpdateRecipe(
